Rename: 코드 리뷰 반영
1. mention Home에서 Common으로 변경
